Rename seed directory
